+++ src/Bundles/Spelling/Exception/SpellingValidationNotSupportedException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace PHPUnuhi\Bundles\Spelling\Exception;
+
+use Exception;
+
+class SpellingValidationNotSupportedException extends Exception
+{
+}